fix(shutdown): release every moving part on the way down
CLI input no longer parks a to_thread worker the loop must join on
exit: one line of stdin now runs on an abandoned daemon thread, so
ctrl+c in terminal mode actually exits.

A running dream is cancelled with the skill instead of writing memory
and holding the background pool while the final save closes it.

Expression.close stops local audio, caption typing and call visuals
before the ports they go through are closed, and brain.shutdown now
closes expression + event streams alongside the stage.

stop_skills waits for the warmup/rhythm loops; OBS disconnect drops
the client so a late caption is a no-op; the stream sentinels are
delivered even on a full queue via fanout.offer.

The discord bot is only asked to leave once: a second SIGTERM reads
as 'now, not cleanly' and would undo the graceful exit.

graceful_timeout is an int, which is what uvicorn takes (pyright).

+++ src/web/server.py
@@ -11,8 +11,9 @@ logger = get_logger("bea.web.server")
 # how long uvicorn waits for open connections (the dashboard's event streams)
 # before it stops waiting and goes on with the shutdown. There used to be no
 # bound at all: ctrl+c with the dashboard open never got past "Waiting for
-# connections to close" until the tab was closed by hand.
-GRACEFUL_TIMEOUT = 5.0
+# connections to close" until the tab was closed by hand. Uvicorn takes this in
+# whole seconds, so it is an int, not a float.
+GRACEFUL_TIMEOUT = 5
 
 
 def begin_shutdown(brain) -> None:
@@ -47,7 +48,7 @@ def begin_shutdown(brain) -> None:
 
 
 async def run_server(brain, host: str = "127.0.0.1", port: int = 8000,
-                     on_shutdown=None, graceful_timeout: float = GRACEFUL_TIMEOUT):
+                     on_shutdown=None, graceful_timeout: int = GRACEFUL_TIMEOUT):
     """Serves the dashboard + brain API.
 
     Binds to loopback by default: no endpoint is authenticated, so exposing the
